# swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"API for Postaway application",
        "title":"Postaway API"
    },
    "servers":[
        {
            "url":"http://localhost:3000"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
            "type": "apiKey",
            "in": "cookie",
            "name": "jwttoken"
            }
        }
    },
    "paths":
...
        "/api/posts": {
            "get": {
                "tags": ["Posts"],
                "summary": "Get User's Posts",
                "description": "Get posts created by the logged-in user",
                "security": [{ "cookieAuth": {} }],
                "responses": {
                "200": {
                    "description": "User's posts fetched successfully"
                }
                }
            },
            "post": {
                "tags": ["Posts"],
                "summary": "Add Post",
                "description": "Create a new post with optional file upload",
                "security": [{ "cookieAuth": {} }],
                "requestBody": {
                "content": {
                    "multipart/form-data": {
                    "schema": {
                        "type": "object",
                        "properties": {
                        "title": { "type": "string" },
                        "description": { "type": "string" },
                        "file": {
                            "type": "string",
                            "format": "binary"
                        }
                        },
                        "required": ["title", "description"]
                    }
                    }
                }
                },
                "responses": {
                "201": {
                    "description": "Post created successfully"
                }
                }
            }
        },
...
    }
}